feat: Implement JsonOverrideProvider
This is pretty raw, I don't like where this storage is in MapDataService but for all intents and purposes of UI development this is working fine.
