+++ common/src/main/java/com/wynntils/services/mapdata/providers/json/JsonProvider.java
@@ -4,49 +4,24 @@
  */
 package com.wynntils.services.mapdata.providers.json;
 
-import com.google.gson.JsonDeserializationContext;
-import com.google.gson.JsonDeserializer;
-import com.google.gson.JsonElement;
-import com.google.gson.JsonObject;
-import com.google.gson.JsonParseException;
-import com.google.gson.JsonSerializationContext;
-import com.google.gson.JsonSerializer;
-import com.google.gson.JsonSyntaxException;
-import com.google.gson.reflect.TypeToken;
 import com.google.gson.stream.MalformedJsonException;
 import com.wynntils.core.WynntilsMod;
 import com.wynntils.core.components.Managers;
 import com.wynntils.core.net.Download;
-import com.wynntils.services.mapdata.attributes.impl.MapAreaAttributesImpl;
-import com.wynntils.services.mapdata.attributes.impl.MapAttributesImpl;
-import com.wynntils.services.mapdata.attributes.impl.MapLocationAttributesImpl;
-import com.wynntils.services.mapdata.attributes.impl.MapPathAttributesImpl;
-import com.wynntils.services.mapdata.attributes.type.MapAreaAttributes;
-import com.wynntils.services.mapdata.attributes.type.MapLocationAttributes;
-import com.wynntils.services.mapdata.attributes.type.MapPathAttributes;
-import com.wynntils.services.mapdata.features.impl.MapAreaImpl;
-import com.wynntils.services.mapdata.features.impl.MapLocationImpl;
-import com.wynntils.services.mapdata.features.impl.MapPathImpl;
 import com.wynntils.services.mapdata.features.type.MapFeature;
-import com.wynntils.services.mapdata.impl.MapCategoryImpl;
-import com.wynntils.services.mapdata.impl.MapIconImpl;
 import com.wynntils.services.mapdata.providers.type.MapDataProvider;
 import com.wynntils.services.mapdata.type.MapCategory;
 import com.wynntils.services.mapdata.type.MapDataProvidedType;
 import com.wynntils.services.mapdata.type.MapIcon;
-import com.wynntils.utils.JsonUtils;
 import java.io.File;
 import java.io.FileReader;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.Reader;
-import java.lang.reflect.Type;
 import java.net.URI;
 import java.nio.charset.StandardCharsets;
-import java.util.Base64;
 import java.util.List;
-import java.util.Objects;
 import java.util.function.BiConsumer;
 import java.util.function.Consumer;
 import java.util.stream.Stream;
@@ -60,7 +35,7 @@ public final class JsonProvider implements MapDataProvider {
     private final List<MapCategory> categories;
     private final List<MapIcon> icons;
 
-    private JsonProvider(int version, JsonFeatures features, List<MapCategory> categories, List<MapIcon> icons) {
+    public JsonProvider(int version, JsonFeatures features, List<MapCategory> categories, List<MapIcon> icons) {
         this.version = version;
         this.features = features;
         this.categories = categories;
@@ -161,190 +136,5 @@ public final class JsonProvider implements MapDataProvider {
     @Override
     public void reloadData() {
         // Json providers do not need to reload data, as the whole object is reloaded by the owner service instead
-    }
-
-    public static final class JsonProviderSerializer implements JsonDeserializer<JsonProvider> {
-        @Override
-        public JsonProvider deserialize(JsonElement json, Type type, JsonDeserializationContext context)
-                throws JsonParseException {
-            JsonObject jsonProvider = json.getAsJsonObject();
-
-            int version = JsonUtils.getNullableJsonInt(jsonProvider, "version");
-            if (version == 0) {
-                throw new JsonParseException("Missing version field in json provider.");
-            }
-            if (version > 1) {
-                throw new JsonParseException("Unsupported version in json provider.");
-            }
-
-            JsonObject featuresObject = JsonUtils.getNullableJsonObject(jsonProvider, "features");
-
-            // Manually construct the JsonFeatures object, so we don't have empty lists
-            JsonFeatures features = new JsonFeatures(
-                    context.deserialize(
-                            JsonUtils.getNullableJsonArray(featuresObject, "locations"),
-                            new TypeToken<List<MapLocationImpl>>() {}.getType()),
-                    context.deserialize(
-                            JsonUtils.getNullableJsonArray(featuresObject, "areas"),
-                            new TypeToken<List<MapAreaImpl>>() {}.getType()),
-                    context.deserialize(
-                            JsonUtils.getNullableJsonArray(featuresObject, "paths"),
-                            new TypeToken<List<MapPathImpl>>() {}.getType()));
-
-            if (!features.validate()) {
-                throw new JsonParseException("Invalid features in json provider.");
-            }
-
-            List<MapCategory> categories = context.deserialize(
-                    JsonUtils.getNullableJsonArray(jsonProvider, "categories"),
-                    new TypeToken<List<MapCategory>>() {}.getType());
-            List<MapIcon> icons = context.deserialize(
-                    JsonUtils.getNullableJsonArray(jsonProvider, "icons"), new TypeToken<List<MapIcon>>() {}.getType());
-
-            return new JsonProvider(version, features, categories, icons);
-        }
-    }
-
-    public static final class JsonCategorySerializer implements JsonDeserializer<MapCategoryImpl> {
-        @Override
-        public MapCategoryImpl deserialize(JsonElement jsonElement, Type type, JsonDeserializationContext context)
-                throws JsonSyntaxException {
-            JsonObject json = jsonElement.getAsJsonObject();
-
-            String id = json.get("id").getAsString();
-            String name = JsonUtils.getNullableJsonString(json, "name");
-            JsonElement attributesJson = json.get("attributes");
-            MapAttributesImpl attributes = attributesJson == null
-                    ? null
-                    : Managers.Json.GSON.fromJson(attributesJson, MapAttributesImpl.class);
-
-            return new MapCategoryImpl(id, name, attributes);
-        }
-    }
-
-    public static final class JsonAttributeSerializer
-            implements JsonDeserializer<MapAttributesImpl>, JsonSerializer<MapAttributesImpl> {
-        @Override
-        public MapAttributesImpl deserialize(JsonElement json, Type type, JsonDeserializationContext context)
-                throws JsonParseException {
-            JsonObject attributesJson = json.getAsJsonObject();
-            MapAttributesImpl attributesObj = context.deserialize(json, MapAttributesImpl.class);
-
-            // We might not want a specific implementation, for example for categories
-            if (type.equals(MapAttributesImpl.class)) {
-                return attributesObj;
-            }
-
-            Type locationType = new TypeToken<MapLocationAttributesImpl>() {}.getType();
-            Type pathType = new TypeToken<MapPathAttributesImpl>() {}.getType();
-            Type areaType = new TypeToken<MapAreaAttributesImpl>() {}.getType();
-
-            if (type.equals(locationType)) {
-                MapLocationAttributes.getUnsupportedAttributes().forEach(invalidAttribute -> {
-                    if (attributesJson.getAsJsonObject().has(invalidAttribute)) {
-                        WynntilsMod.warn("Unsupported attribute set for location: " + invalidAttribute);
-                    }
-                });
-
-                return new MapLocationAttributesImpl(attributesObj);
-            }
-
-            if (type.equals(areaType)) {
-                MapAreaAttributes.getUnsupportedAttributes().forEach(invalidAttribute -> {
-                    if (attributesJson.getAsJsonObject().has(invalidAttribute)) {
-                        WynntilsMod.warn("Unsupported attribute set for area: " + invalidAttribute);
-                    }
-                });
-
-                return new MapAreaAttributesImpl(attributesObj);
-            }
-
-            if (type.equals(pathType)) {
-                MapPathAttributes.getUnsupportedAttributes().forEach(invalidAttribute -> {
-                    if (attributesJson.getAsJsonObject().has(invalidAttribute)) {
-                        WynntilsMod.warn("Unsupported attribute set for path: " + invalidAttribute);
-                    }
-                });
-
-                return new MapPathAttributesImpl(attributesObj);
-            }
-
-            throw new JsonParseException("Attribute type is not location, path or area");
-        }
-
-        @Override
-        public JsonElement serialize(MapAttributesImpl mapAttributes, Type type, JsonSerializationContext context) {
-            // For the base class, we just serialize the attributes as is
-            if (type.equals(MapAttributesImpl.class)) {
-                return Managers.Json.GSON.toJsonTree(mapAttributes);
-            }
-
-            // For the specific implementations, we need to serialize the attributes that make sense,
-            // explicitly skipping null and unsupported attributes
-            JsonObject attributesJson = Managers.Json.GSON
-                    .toJsonTree(mapAttributes, MapAttributesImpl.class)
-                    .getAsJsonObject();
-            JsonObject result = new JsonObject();
-
-            attributesJson.entrySet().forEach(entry -> {
-                String attribute = entry.getKey();
-                JsonElement value = entry.getValue();
-                if (value != null) {
-                    result.add(attribute, value);
-                }
-            });
-
-            if (type.equals(MapLocationAttributesImpl.class)) {
-                MapLocationAttributes.getUnsupportedAttributes().forEach(result::remove);
-            } else if (type.equals(MapPathAttributesImpl.class)) {
-                MapPathAttributes.getUnsupportedAttributes().forEach(result::remove);
-            } else if (type.equals(MapAreaAttributesImpl.class)) {
-                MapAreaAttributes.getUnsupportedAttributes().forEach(result::remove);
-            } else {
-                throw new JsonParseException("Attribute type is not location, path or area, nor the base class");
-            }
-
-            return result;
-        }
-    }
-
-    public static final class JsonIconSerializer implements JsonDeserializer<MapIconImpl>, JsonSerializer<MapIconImpl> {
-        @Override
-        public MapIconImpl deserialize(JsonElement jsonElement, Type jsonType, JsonDeserializationContext context)
-                throws JsonParseException {
-            JsonObject json = jsonElement.getAsJsonObject();
-
-            String id = json.get("id").getAsString();
-            String base64Texture = json.get("texture").getAsString();
-
-            byte[] texture = Base64.getDecoder().decode(base64Texture);
-
-            try {
-                return new MapIconImpl(id, texture);
-            } catch (IOException e) {
-                WynntilsMod.warn("Bad icon texture for " + id, e);
-                return null;
-            }
-        }
-
-        @Override
-        public JsonElement serialize(MapIconImpl mapIcon, Type type, JsonSerializationContext context) {
-            JsonObject json = new JsonObject();
-            json.addProperty("id", mapIcon.getIconId());
-            json.addProperty("texture", Base64.getEncoder().encodeToString(mapIcon.getTextureBytes()));
-            return json;
-        }
-    }
-
-    private record JsonFeatures(List<MapLocationImpl> locations, List<MapAreaImpl> areas, List<MapPathImpl> paths) {
-        public Stream<MapFeature> stream() {
-            return Stream.of(locations, areas, paths).filter(Objects::nonNull).flatMap(List::stream);
-        }
-
-        public boolean validate() {
-            return locations.stream().allMatch(MapLocationImpl::validate)
-                    && areas.stream().allMatch(MapAreaImpl::validate)
-                    && paths.stream().allMatch(MapPathImpl::validate);
-        }
     }
 }
